Moved Ray tracing engine to Node system fixed warnings Improved display of Count timers and GPU timers

+++ GraphicsEngine/Source/Core/Core/Platform/Windows/WindowPlatformMisc.cpp
@@ -42,6 +42,6 @@ PlatformMemoryInfo WindowPlatformMisc::GetMemoryInfo()
 void WindowPlatformMisc::SetCurrnetThreadAffinity(int core)
 {
 	int MAsk = 1 << core;
-	int value = SetThreadAffinityMask(GetCurrentThread(), MAsk);
+	int value = (int)SetThreadAffinityMask(GetCurrentThread(), MAsk);
 	ensure(value != 0);
 }
