+++ GraphicsEngine/Source/D3D12RHI/RHI/RenderAPIs/D3D12/D3D12TimeManager.h
@@ -71,7 +71,7 @@ private:
 		std::string name = "";
 		std::vector<D3D12Query*> TimerQueries;
 		float TotalTime = 0.0f;
-		void Resolve(float freqnecy);
+		void Resolve(UINT64 freqnecy);
 	};
 	UINT64 StartTimeStamp = 0;
 	std::vector<TimerQ> TimerQueries;
